PP-8940 Rename controllers and tests + clean up
- For the new  amount and reference pages
  - Merge everything into 1 controller:
    - New - amount.controller.js
    - New - reference.controller.js
- Update tests
    - New - amount.controller.test.js
    - New - reference.controller.test.js
- Remove the `get-index.controller.js` and `post-index.controller.js` files.
- Update the routes.js
