--- app/payment-link-v2/reference/index.js
@@ -1,4 +1,0 @@
-'use strict'
-
-module.exports.index = require('./get-index.controller')
-module.exports.postIndex = require('./post-index.controller')
